refactor: renamed android/index.ts to android/androidSetup.ts
