+++ src/commands/android/androidSetup.ts
